@@ -33,7 +33,7 @@ setup(
     license='ZPL 2.1',
     description='zope.sendmail integration for Zope.',
     author='Zope Foundation and Contributors',
-    author_email='zope-dev@zope.org',
+    author_email='zope-dev@zope.dev',
     long_description=_read('README.rst') + '\n' + _read('CHANGES.rst'),
     packages=find_packages('src'),
     namespace_packages=['Products'],
@@ -42,16 +42,11 @@ setup(
         'Development Status :: 6 - Mature',
         'Environment :: Web Environment',
         'Framework :: Zope',
-        'Framework :: Zope :: 4',
         'Framework :: Zope :: 5',
         'License :: OSI Approved :: Zope Public License',
         'Operating System :: OS Independent',
         'Programming Language :: Python',
-        'Programming Language :: Python :: 2',
-        'Programming Language :: Python :: 2.7',
         'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.5',
-        'Programming Language :: Python :: 3.6',
         'Programming Language :: Python :: 3.7',
         'Programming Language :: Python :: 3.8',
         'Programming Language :: Python :: 3.9',
@@ -60,10 +55,9 @@ setup(
         'Programming Language :: Python :: Implementation :: CPython',
         'Topic :: Communications :: Email',
     ],
-    python_requires='>=2.7,!=3.0.*,!=3.1.*,!=3.2.*,!=3.3.*,!=3.4.*',
+    python_requires='>=3.7',
     install_requires=[
         'setuptools',
-        'six',
         'AccessControl',
         'Acquisition',
         'DateTime',
