@@ -13,8 +13,6 @@
 """MailHost export / import support.
 """
 
-import six
-
 from zope.component import adapts
 
 from Products.GenericSetup.interfaces import ISetupEnviron
@@ -68,13 +66,8 @@ class MailHostXMLAdapter(XMLAdapterBase):
         self.context.smtp_port = int(node.getAttribute('smtp_port'))
         smtp_uid = node.getAttribute('smtp_uid')
         smtp_pwd = node.getAttribute('smtp_pwd')
-        if six.PY2:
-            # ??? Why?
-            self.context.smtp_uid = smtp_uid.encode('utf-8')
-            self.context.smtp_pwd = smtp_pwd.encode('utf-8')
-        else:
-            self.context.smtp_uid = smtp_uid
-            self.context.smtp_pwd = smtp_pwd
+        self.context.smtp_uid = smtp_uid
+        self.context.smtp_pwd = smtp_pwd
 
         # Older MH instances won't have 'smtp_queue' in instance dict
         if 'smtp_queue' in self.context.__dict__:
